requirements.txt
streamlit>=1.37.0
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0
tldextract>=5.1.2
validators>=0.33.0
python-slugify>=8.0.4
